requirements.txt
markdownify
readability-lxml
trafilatura; python_version > "2.7"
